edited gradle version and package name to run it on android phone

+++ lib/models/todo_model.dart
@@ -6,7 +6,7 @@ class Todo {
   String nickname;
   int age;
   bool isSingle;
-  double happinessLevel;
+  int happinessLevel;
   String radioMotto;
   String superpower;
